Move rustc-specific entrypoint to the `rustc` module
